Pass test suite for FCC tribute page
This commit adds the image responsiveness required to pass the FCC test
suite for the tribute page challenge.  Page layout, fonts, and color
schemes were updated to match the Figma design document created prior to
the challenge.  A pdf copy of the design document can be found in the
figma subdirectory for the project.