Pass test suite for FCC survey page
This commit adds the form input types required to pass the FCC test
suite for the survey page challenge.  A second stylesheet refactor was
required to accomodate the fieldset, select, and radio/checkbox input
tags.  A pdf copy of the Figma design document for the page has been
included in the figma subdirectory.